backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "roofee-backend"
version = "0.1.0"
description = "FastAPI backend for Roofee."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "imagecodecs>=2024.9.22",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "opencv-python-headless>=4.10.0",
  "pillow>=10.4.0",
  "python-multipart>=0.0.9",
  "tifffile>=2024.8.30",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "shapely>=2.0.0",
  "trimesh>=4.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0"
]
vision = [
  "huggingface-hub>=0.24.0",
  "ultralytics>=8.3.0"
]
rid = [
  "numpy<2",
  "segmentation-models==1.0.1",
  "tensorflow==2.10.0; python_version < '3.11'"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
